server/requirements.txt
fastapi
uvicorn
minio
pymongo
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]